4.21* blogilistan laajennus, osa 10

+++ blogilista/controllers/blogs.js
@@ -72,12 +72,28 @@ blogsRouter.put('/:id', async (request, response) => {
 
 blogsRouter.delete('/:id', async (request, response) => {
     try {
+        const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+        if (!request.token || !decodedToken.id) {
+            return response.status(401).json({ error: 'token missing or invalid' })
+        }
+        const user = await User.findById(decodedToken.id)
+        const blog = await Blog.findById(request.params.id)
+        console.log(user.id, String(blog.user[0]))
+        if (user.id !== String(blog.user[0])){
+            return response.status(401).json({ error: 'wrong user' })
+        }
+
         await Blog.findByIdAndRemove(request.params.id)
 
         response.status(204).end()
     } catch (exception) {
-        console.log(exception)
-        response.status(400).send({ error: 'malformatted id' })
+        if (exception.name === 'JsonWebTokenError' ) {
+            response.status(401).json({ error: exception.message })
+        } else {
+            console.log(exception)
+            response.status(400).send({ error: 'malformatted id' })
+        }
     }
 })
